refactor(types, validation): move type aliases and data validation to their modules
