+++ src/llms/client.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import json
 from typing import TYPE_CHECKING
 
 import structlog
@@ -12,22 +11,19 @@ from litellm import (
     supports_response_schema,
     validate_environment,
 )
-from pydantic import BaseModel, ValidationError
+from pydantic import BaseModel
 from tqdm import tqdm
 
+from src.data.data_validators import validate_content_against_schema
 from src.utils import constants as const
 from src.utils import read_config
 
 if TYPE_CHECKING:
-    from typing import Any, Optional, Sequence, TypeVar
+    from typing import Any, Optional, Sequence
 
     from structlog.stdlib import BoundLogger
 
-    from src.utils.aliases import PathLike
-
-    ModelT = TypeVar("ModelT", bound=BaseModel)  # subclass of BaseModel
-    ResponseType = str | ModelT | dict[str, Any]
-    BatchResponseType = list[ResponseType]
+    from src.utils.types import BatchResponseType, ModelT, PathLike, ResponseType
 
 # Set up logging
 logger: BoundLogger = structlog.getLogger(__name__)
@@ -305,119 +301,3 @@ def process_llm_responses(
             processed_responses.append(None)
 
     return processed_responses
-
-
-def validate_content_against_schema(content: Any, schema: type[ModelT]) -> ModelT:
-    """Validate the content against the schema.
-
-    Args:
-        content: The content to validate (string, dict, or model instance)
-        schema: The schema to validate against
-
-    Returns:
-        The content parsed and validated against the schema
-
-    Raises:
-        ValueError: If the content is not a string or dictionary
-        json.JSONDecodeError: If the content is not valid JSON
-        ValidationError: If the content does not match the schema
-    """
-    if content is None:
-        logger.exception("Content is None")
-        raise ValueError("Content is None")
-    try:
-        # if content is already schema, return it
-        if isinstance(content, schema):
-            return content
-        elif isinstance(content, dict):
-            logger.debug("Validating dictionary against schema")
-            return schema.model_validate(content)
-
-        elif isinstance(content, (str, bytes, bytearray)):
-            try:
-                logger.debug("Validating JSON string against schema")
-                return schema.model_validate_json(content)
-            except ValidationError as e:
-                # Try parsing JSON first then validate
-                logger.warning(
-                    "Direct validation failed, parsing JSON first", error=str(e)
-                )
-                parsed_json = json.loads(content)
-                return schema.model_validate(parsed_json)
-
-        # Fallback for other types
-        else:
-            logger.warning(f"Unexpected content type: {type(content)}")
-            content_str = str(content)
-            return schema.model_validate_json(content_str)
-
-    except json.JSONDecodeError as json_error:
-        content_str = str(content)
-        snippet = content_str[:100] + "..." if len(content_str) > 100 else content_str
-        logger.exception("JSON decode error", error=str(json_error), content=snippet)
-
-        # Try to fix incomplete JSON
-        fixed_content = _attempt_fix_incomplete_json(str(content))
-        fixed_snippet = (
-            fixed_content[:100] + "..." if len(fixed_content) > 100 else fixed_content
-        )
-        logger.debug("Attempting with fixed JSON", fixed_content=fixed_snippet)
-
-        try:
-            return schema.model_validate_json(fixed_content)
-        except Exception as fix_error:
-            logger.exception("Error validating fixed JSON", error=str(fix_error))
-            raise fix_error
-
-    except ValidationError as validation_error:
-        logger.exception("Error validating content with schema")
-        raise validation_error
-
-    except Exception as e:
-        logger.exception("Unexpected error validating content")
-        raise e
-
-
-def _attempt_fix_incomplete_json(content: str) -> str:
-    """Attempt to fix incomplete JSON by closing open brackets and quotes.
-
-    Args:
-        content: The incomplete JSON string
-
-    Returns:
-        The fixed JSON string
-    """
-    # Count open brackets and quotes
-    open_braces = content.count("{") - content.count("}")
-    open_brackets = content.count("[") - content.count("]")
-    open_quotes = content.count('"') % 2
-
-    # Fix open braces and brackets
-    fixed_content = content
-
-    # Handle trailing commas in objects or arrays
-    if fixed_content.rstrip().endswith(","):
-        fixed_content = fixed_content.rstrip()[:-1]
-
-    # First fix quotes if needed
-    if open_quotes > 0:
-        # Try to find the last property name or value
-        last_quote = content.rfind('"')
-        if last_quote > 0:
-            # Check if we're in a property name or value
-            prev_colon = content.rfind(":", 0, last_quote)
-            prev_comma = content.rfind(",", 0, last_quote)
-
-            if prev_colon > prev_comma:  # We're in a value
-                fixed_content = content + '"'
-            else:  # We're in a property name
-                fixed_content = content + '":""'
-
-    # Then close brackets and braces
-    if open_brackets > 0:
-        fixed_content += "]" * open_brackets
-
-    if open_braces > 0:
-        fixed_content += "}" * open_braces
-
-    return fixed_content
